+++ CHANGELOG.md
@@ -61,3 +61,15 @@ Change String angsuran perbulan & jatuh tempo into widget on DSCard
 ## [0.0.1+15] - 16 February 2021.
 
 Add General DS Button, Add Card for Product Information
+
+## [0.0.1+16] - 16 February 2021.
+
+Change Image Asset to Network on Card Product Information
+
+## [0.0.1+17] - 16 February 2021.
+
+Add font color for General DS Button
+
+## [0.0.1+18] - 17 February 2021.
+
+Add General DS Ghost Button